Reverting to float32 calculations for imagestats since multiple algorithms assume this data type in their calculations.
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/imagestats@920 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/__init__.py
@@ -155,7 +155,7 @@ class ImageStats:
 
             if (self.fields.find('median') != -1):
                 # Compute Median Value
-                _binSum = N.cumsum(_bins).astype(N.float64)
+                _binSum = N.cumsum(_bins).astype(N.float32)
                 _binSum = _binSum/_binSum[-1]
                 _lo = N.where(_binSum >= 0.5)[0][0]
                 _hi = _lo + 1
